requirements.txt
asgiref==3.2.3
backcall==0.1.0
decorator==4.4.1
Django==3.0.3
django-debug-toolbar==2.2
ipython==7.12.0
ipython-genutils==0.2.0
jedi==0.16.0
mysqlclient==1.4.6
numpy==1.18.1
parso==0.6.1
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.3
ptyprocess==0.6.0
Pygments==2.5.2
pytz==2019.3
readline==6.2.4.1
scipy==1.4.1
six==1.14.0
sqlparse==0.3.0
traitlets==4.3.3
wcwidth==0.1.8
